opa-react: move sdk call into useAuthz hook
Since there are multiple uses of <Authz> or useAuthz, we cannot set these things with
the AuthzProvider. AuthzProvider is merely providing the SDK instance, and some defaults
for path and input.
