+++ client/react/packages/opa-react/src/use-authz.ts
@@ -1,6 +1,7 @@
-import { useContext, useEffect, useMemo } from "react";
+import { useContext, useMemo } from "react";
 import { AuthzContext } from "./opa-provider";
 import { Input, Result } from "@styra/opa";
+import merge from "lodash/merge";
 
 export interface UseAuthzResult {
   result: Result | undefined;
@@ -49,19 +50,21 @@ export default function useAuthz(path?: string, input?: Input): UseAuthzResult {
   if (context === null) {
     throw Error("Authz/useAuthz can only be used inside an AuthzProvider");
   }
-  const { result, setInput, setPath } = context;
-  useEffect(() => {
-    setInput(input);
-  }, [input]);
-
-  useEffect(() => {
-    if (path) setPath(path);
-  }, [path]);
+  const { sdk, defaultPath, defaultInput } = context;
 
   return useMemo<UseAuthzResult>(() => {
+    const p = path ?? defaultPath;
+    const i = mergeInput(input, defaultInput);
+    const result = sdk.evaluate<Input, Result>(p, i);
     return {
       result,
       isLoading: result === undefined,
     };
-  }, [result, input, path, setInput]);
+  }, [sdk, path, input, defaultPath, defaultInput]);
+}
+
+function mergeInput(input?: Input, defaultInput?: Input): Input | undefined {
+  if (!input) return defaultInput;
+  if (!defaultInput) return input;
+  return merge(input, defaultInput);
 }
